Extract createCachedSelector logic from index.js
