+++ src/createStructuredCachedSelector.js
@@ -1,5 +1,5 @@
 import {createStructuredSelector} from 'reselect';
-import createCachedSelector from './index';
+import createCachedSelector from './createCachedSelector';
 
 function createStructuredCachedSelector(selectors) {
   return createStructuredSelector(selectors, createCachedSelector);
